refactor(encore): split dispatch.ts into per-kind handler modules
dispatch.ts was 827 lines holding 11 zod schemas, 11 handler bodies,
shared helpers, and the lock + router. Each handler now lives in its
own file under `server/encore/handlers/`, with cross-handler helpers
(`loadDsl`, `loadCycle`, `persistAndReconcile`, `assertKnownTargetAndStep`,
`coerceDefinitionToObject`, `formatZodError`, `workspaceRelativePath`,
`EncoreError`, dispatch types) in `handlers/shared.ts`.

dispatch.ts is now just the lock envelope, the kind→handler routing,
and the `safeParse` wrapper that converts zod failures into 400s. No
behavior change; public exports (`dispatch`, `EncoreError`,
`EncoreDispatchBody`, `EncoreDispatchResult`) keep the same names so
`server/api/routes/encore.ts` and the dispatch/reconcile tests don't
need updates. Addresses P0.1 of the Codex review.
